Move SenML cbor class in their own package.
