+++ leshan-core/src/main/java/org/eclipse/leshan/senml/cbor/SenMLCborException.java
@@ -12,15 +12,21 @@
  *     Boya Zhang - initial API and implementation
  *******************************************************************************/
 
-package org.eclipse.leshan.senml;
+package org.eclipse.leshan.senml.cbor;
 
 /**
- * Helper for encoding/decoding SenML CBOR format
+ * Exception thrown in case of Cbor encoding error
  */
-public class SenMLCbor {
-    private static final SenMLCborPackSerDes serDes = new SenMLCborPackSerDes();
+public class SenMLCborException extends Exception {
 
-    public static byte[] toSenMLCbor(SenMLPack pack) throws SenMLCborException {
-        return serDes.serializeToCbor(pack);
+    private static final long serialVersionUID = 1L;
+
+    public SenMLCborException(String message) {
+        super(message);
     }
+
+    public SenMLCborException(String message, Exception cause) {
+        super(message, cause);
+    }
+
 }
